pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lewm"
version = "0.1.0"
description = "LeWorldModel research package"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "LeWM Authors" },
]
dependencies = [
  "einops",
  "hydra-core",
  "lightning",
  "numpy",
  "omegaconf",
  "scikit-learn",
  "stable-pretraining",
  "stable-worldmodel[train,env]",
  "torch",
  "torchvision",
]

[project.urls]
Source = "https://github.com/lucas-maes/le-wm"
Issues = "https://github.com/lucas-maes/le-wm/issues"

[project.scripts]
lewm-train = "lewm.train_app:main"
lewm-eval = "lewm.eval_app:main"

[project.optional-dependencies]
dev = []

[tool.uv]
package = true

[tool.hatch.version]
path = "src/lewm/__init__.py"

[tool.hatch.build]
exclude = [
  ".worktrees/**",
]

[tool.hatch.build.targets.wheel]
packages = ["src/lewm"]
only-packages = true

[tool.hatch.build.targets.wheel.force-include]
"src/lewm/config/train" = "lewm/config/train"
"src/lewm/config/eval" = "lewm/config/eval"

[tool.hatch.build.targets.sdist]
include = [
  "/src",
  "/tests",
  "/README.md",
  "/CLAUDE.md",
  "/LICENSE",
]

[tool.hatch.build.targets.sdist.force-include]
"src/lewm/config/train" = "src/lewm/config/train"
"src/lewm/config/eval" = "src/lewm/config/eval"

[tool.hatch.envs.default]
dependencies = []

[tool.hatch.envs.default.scripts]
package-layout = "python tests/test_package_layout.py"
shape-tests = "python tests/test_jepa_shapes.py"
test = "python -m unittest discover -s tests -p 'test_*.py'"
